+++ 1. Suit 'em up Game Development/SuitEmUp/SuitEmUp/SuperEnemy.h
@@ -12,7 +12,7 @@ public:
 	~SuperEnemy();
 	bool Update(float deltatime);
 	bool GetType();
-	int Damaged(int playerdmg);
+	float Damaged(float playerdmg);
 	float GetDamage();
 	float GetRotation();
 
@@ -20,5 +20,5 @@ protected:
 	float rotation;
 	float m_damage;
 	float m_cooldown;
-	int m_hp;
+	float m_hp;
 };
